+++ src/zabbix_server/trapper/active.c
@@ -95,8 +95,8 @@ static void	db_register_host(const char *host, const char *ip, unsigned short po
 	{
 		zbx_db_begin();
 
-		zbx_autoreg_update_host(0, host, p_ip, p_dns, port, connection_type, host_metadata, (int)flag, now,
-				events_cbs);
+		zbx_autoreg_update_host(0, host, p_ip, p_dns, port, connection_type, host_metadata,
+				(unsigned short)flag, now, events_cbs);
 	}
 	while (ZBX_DB_DOWN == zbx_db_commit());
 }
